+++ CHANGELOG.md
@@ -3,7 +3,12 @@
 </a>
 
 # Changelog
-## Latest Version ## v2.2.0 (07/30/26)
+## Latest Version ## v2.3.0 (08/13/26)
+- Added Support for eRaty via Hosted Payment Page (HPP)
+- Fixed Apm(blik and payu) issue for upgraded JS version
+- Improved compatibility with some hosting environments when using GPAPI & HPP
+
+## v2.2.0 (07/30/26)
 - Added Support for Visa Installments Payment Plans
 - Fixed HPP checkout ignoring "Authorize" setting
 - Fixed HPP visa installments error for non-UK&I merchants